+++ SolastaCommunityExpansion/Settings.cs
@@ -220,7 +220,7 @@ namespace SolastaCommunityExpansion
         public bool EnableHotkeysToToggleHud { get; set; }
         public bool InvertAltBehaviorOnTooltips { get; set; }
         public bool ShowCraftingRecipeInDetailedTooltips { get; set; }
-        public bool AdditionalIconsOnLevelMap { get; set; }
+        public bool EnableAdditionalIconsOnLevelMap { get; set; }
 
 
         // Adventure Log
